requirements.txt
aiogram==3.13.1
beautifulsoup4==4.12.3
lxml==5.2.0
python-dotenv==1.0.1
requests==2.32.3
user-agent==0.1.10
